Optimize carbon structure
